requirements.txt
asgiref==3.10.0
brotli==1.2.0
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
Django==5.2.8
django-rest-framework==0.1.0
djangorestframework==3.16.1
fonttools==4.60.1
pillow==12.0.0
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
reportlab==4.4.4
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==66.0
webencodings==0.5.1
zopfli==0.4.0
